move data module to task root (out of control)
